# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentación",
        "description": "Documentación de la API usando L5 Swagger",
        "contact": {
            "email": "soporte@tudominio.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor API"
        }
    ],
    "paths": {
        "/api/auth": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Autenticar usuario y generar token JWT",
                "operationId": "81d1cbd5fbb95deefc65bece9802eb46",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token JWT generado correctamente"
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    }
                }
            }
        },
        "/api/tasks/send-soap": {
            "post": {
                "tags": [
                    "Soap Services"
                ],
                "summary": "Enviar tareas por SOAP",
                "operationId": "1523ec4b5dabbd7bfb601c37e7e469a5",
                "responses": {
                    "200": {
                        "description": "Tareas enviadas exitosamente"
                    },
                    "500": {
                        "description": "Error al enviar tareas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Obtener lista de tareas con paginación",
                "operationId": "471d938a3bfaeb6fbd981d7c3e2728b0",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "deleted",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "created_at",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-07-15"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de tareas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Crear nueva tarea",
                "operationId": "f803ebdb2300acdb7e93bbe8d0b9a2e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "priority_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Nueva tarea"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción opcional"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "pending"
                                    },
                                    "priority_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "image": {
                                        "description": "Imagen de la tarea",
                                        "type": "file"
                                    },
                                    "reminder_time": {
                                        "description": "Fecha y hora para el recordatorio (formato Y-m-d H:i:s)",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-07-15 14:00:00"
                                    },
                                    "reminder_before": {
                                        "description": "Cantidad de tiempo antes del recordatorio",
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "reminder_unit": {
                                        "description": "Unidad de tiempo para el recordatorio (minutes, hours, days)",
                                        "type": "string",
                                        "enum": [
                                            "minutes",
                                            "hours",
                                            "days"
                                        ],
                                        "example": "minutes"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tarea creada correctamente"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tasks/{id}": {
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Eliminar una tarea",
                "description": "Realiza un soft delete de la tarea especificada por su ID.",
                "operationId": "ad582480cd57c2a4e3b6a426943eeca5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tarea a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarea eliminada correctamente"
                    },
                    "404": {
                        "description": "Tarea no encontrada"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Actualizar una tarea existente",
                "description": "Actualiza los datos de una tarea específica por su ID.",
                "operationId": "f70bc5dad1f606e2b2d8b3835be2a3ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tarea a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Tarea actualizada"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Nueva descripción de la tarea"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "in_progress",
                                            "completed"
                                        ],
                                        "example": "in_progress"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "priority_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tarea actualizada correctamente"
                    },
                    "405": {
                        "description": "Tarea actualizada correctamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tasks/export-xml": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Exportar tareas a XML",
                "operationId": "0383f194c9723296f91c29db5d95ed5e",
                "responses": {
                    "200": {
                        "description": "Archivo XML con tareas exportadas"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tasks/restore-xml": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Restaurar tareas desde archivo XML",
                "operationId": "a2c568e87bd471ea14e2f2135038e77a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "Archivo XML para restaurar tareas",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tareas restauradas correctamente"
                    },
                    "400": {
                        "description": "Archivo inválido"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Soap Services",
            "description": "Soap Services"
        },
        {
            "name": "Tasks",
            "description": "Tasks"
        }
    ]
}